Moving the IES adapter code into the XML adapter to make the gbXML adapter accessible through the adapter components
