+++ XML_Adapter/CRUD/Delete.cs
@@ -1,0 +1,15 @@
+﻿using BH.Adapter;
+
+namespace BH.Adapter.gbXML
+{
+    public partial class XMLAdapter : BHoMAdapter
+    {
+        /***************************************************/
+        /**** Public Methods                            ****/
+        /***************************************************/
+
+
+
+        /***************************************************/
+    }
+}
